Barnes Points (#1)
Implements a two versions of the Barnes Points system with support for Combined points, as well as the break down of Women's/Men's Sweep/Sculling points.

Traditional Barnes System: Events of the same boat class across all rower experience level are weighted the same. For example, a varsity 8+ and a novice 8+ will both be worth 30 points.

Weighted Barnes System: Varsity events are worth the full 100% of points for the event, Junior/2V events are 80%, Novice/Freshman/3V are worth 60%. This is the system used by Midwest Scholastic Rowing Association and Scholastic Rowing Association of Michigan.